+++ TimeReg/ViewModels/UserTimePerProjectViewModel.cs
@@ -12,14 +12,9 @@ namespace TimeReg.ViewModels
         public string UserName { get; set; }
         public string NK_ZId { get; set; }
         public string Name { get; set; }
-        public Nullable<int> timeSum { get; set; }
+        public Nullable<double> timeSum { get; set; }
         public int FK_ProjectId { get; set; }
         public List<WeeklyTimeViewModel> WeeklyTimePerProject { get; set; }
-
-
-
     }
 
-  
-    
 }
