+++ TimeReg/ViewModels/ViewModelTools/WeeklyTimeViewModel.cs
@@ -9,7 +9,7 @@ namespace TimeReg.ViewModels.ViewModelTools
     {
         public int Week { get; set; }
         public int Year { get; set; }
-        public int TotalTime { get; set; }
+        public double TotalTime { get; set; }
         public string Name { get; set; }
